Now, we should be able to load special dragons in tournament, using Builds.
Former-commit-id: 47924924a11fe166707e209f3bc4c86920bba776

+++ Assets/Tools/LiveEvents/Editor/LeagueLeaderboardGeneratorEditor.cs
@@ -181,7 +181,7 @@ public class LeagueLeaderboardGeneratorEditor : EditorWindow {
 			newRecord.score = (long)score.GetRandom();
 
 			// Dragon level
-			newRecord.build.level = (uint)dragonLevel.GetRandom();
+			newRecord.build.level = dragonLevel.GetRandom();
 
 			// Dragon
 			newRecord.build.dragon = specialDragonDefs.GetRandomValue().sku;
